% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{AKSProvider-class}
\alias{AKSProvider-class}
\alias{.AKSProvider}
\title{The Azure Kubenetes service provider}
\description{
The Azure Kubenetes service provider for the \code{DockerParallel} package. Please
call \code{AKSProvider()} to create the provider.
}

