% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKSProvider-method.R
\name{AKSProvider}
\alias{AKSProvider}
\title{Create an AKS provider object}
\usage{
AKSProvider(
  k8sCluster = NULL,
  poolName = "akspool",
  AKSName = Sys.getenv("AZ_AKS_NAME"),
  resourceGroupName = Sys.getenv("AZ_RESOURCE_GROUP_NAME"),
  subscriptionName = Sys.getenv("AZ_SUBSCRIPTION"),
  tenant = Sys.getenv("AZ_TENANT"),
  tenantSelection = character(0),
  autoDelete = NULL
)
}
\arguments{
\item{k8sCluster}{KubernetesCluster, the K8S cluster object from the package \code{AzureContainers}}

\item{poolName}{Character(1), the name of the node pool. This parameter
will only be used when the k8s cluster does not exist and need to be created.}

\item{AKSName}{Character(1), the Azure Kubernetes service name}

\item{resourceGroupName}{Character(1), the resource group name. If empty, the function
will print a menu and ask you to choose one.}

\item{subscriptionName}{Character(1), the subscription name. If empty, the function
will print a menu and ask you to choose one.}

\item{tenant}{Character(1), the tenant you want to use. This argument will be passed to
\code{AzureRMR::get_azure_login} to get the ARM login client.}

\item{tenantSelection}{Character(1), if you have multiple logins for a given tenant,
which one to use. This can be a number, or the input MD5 hash of the token used for the login
(You can find them via \code{AzureRMR::list_azure_logins()}). If not supplied,
we will print a menu and ask you to choose a login.}

\item{autoDelete}{Logical(1), whether the kubernetes cluster should be deleted when the
\code{DockerCluster} object has been removed from the workspace? If \code{NULL}, the kubernetes cluster
will only be removed if it is created by this function.}
}
\value{
An \code{AKSProvider} reference object
}
\description{
Create an AKS provider object, you must have a login cache made by
\code{AzureRMR::create_azure_login} before you call this function.
}
\examples{
provider <- AKSProvider(tenant = "Your tenant ID")
}
