% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor.R
\name{cleanupDockerCluster,AKSProvider-method}
\alias{cleanupDockerCluster,AKSProvider-method}
\title{Cleanup the cluster}
\usage{
\S4method{cleanupDockerCluster}{AKSProvider}(provider, cluster, verbose)
}
\value{
No return value
}
\description{
Cleanup the cluster
}
