% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKSProvider-adaptor.R
\name{getDockerServerIp,AKSProvider-method}
\alias{getDockerServerIp,AKSProvider-method}
\title{Get the instance public/private IPs}
\usage{
\S4method{getDockerServerIp}{AKSProvider}(provider, cluster, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A list with the element name \code{publicIp}, \code{publicPort}, \code{privateIp} and \code{privatePort}
}
\description{
Get the instance public/private IPs
}
