% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKSProvider-adaptor.R
\name{getDockerWorkerNumbers,AKSProvider-method}
\alias{getDockerWorkerNumbers,AKSProvider-method}
\title{Get the worker number}
\usage{
\S4method{getDockerWorkerNumbers}{AKSProvider}(provider, cluster, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A list with running and initializing worker numbers
}
\description{
Get the worker number
}
