% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKSProvider-adaptor.R
\name{getExportedNames,AKSProvider-method}
\alias{getExportedNames,AKSProvider-method}
\title{Get the extra function defined for the AKSProvider}
\usage{
\S4method{getExportedNames}{AKSProvider}(x)
}
\arguments{
\item{x}{A cloud provider or container object}
}
\value{
A character vector
}
\description{
Get the extra function defined for the AKSProvider
}
