% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKSProvider-adaptor.R
\name{getExportedObject,AKSProvider-method}
\alias{getExportedObject,AKSProvider-method}
\title{Get the extra function defined for the AKSProvider}
\usage{
\S4method{getExportedObject}{AKSProvider}(x, name)
}
\arguments{
\item{x}{A cloud provider or container object}

\item{name}{The name of the exported object}
}
\value{
A function
}
\description{
Get the extra function defined for the AKSProvider
}
