% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor.R
\name{initializeProvider,AKSProvider-method}
\alias{initializeProvider,AKSProvider-method}
\title{Initialize the provider}
\usage{
\S4method{initializeProvider}{AKSProvider}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Initialize the provider
}
