% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKSProvider-adaptor.R
\name{reconnectDockerCluster,AKSProvider-method}
\alias{reconnectDockerCluster,AKSProvider-method}
\title{Reconnect to the cluster}
\usage{
\S4method{reconnectDockerCluster}{AKSProvider}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Reconnect to the cluster with the same job queue name.
}
