% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor.R
\name{setDockerWorkerNumber,AKSProvider-method}
\alias{setDockerWorkerNumber,AKSProvider-method}
\title{Run the worker containers}
\usage{
\S4method{setDockerWorkerNumber}{AKSProvider}(
  provider,
  cluster,
  container,
  hardware,
  workerNumber,
  verbose = 0L
)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object.}

\item{hardware}{S4 \code{DockerHardware} Object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Run the worker containers
}
