% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor.R
\name{stopDockerServer,AKSProvider-method}
\alias{stopDockerServer,AKSProvider-method}
\title{Stop the server}
\usage{
\S4method{stopDockerServer}{AKSProvider}(provider, cluster, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Stop the server
}
